requirements.txt
beautifulsoup4
faiss-cpu
numpy
python-dotenv
openai
cohere
